requirements.txt
langchain
langchain_community
langchain_core
langchain_google_genai
streamlit
google-generativeai
